% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb2molecule.R
\name{.h2m}
\alias{.h2m}
\title{how to make herb to molecule}
\usage{
.h2m(x, type = "latin")
}
\arguments{
\item{x}{an variable}
}
\value{
molecule_name
}
\description{
how to make herb to molecule
}
\examples{
.h2m(x='Ziziphi Spinosae Semen',type='latin')
.h2m(x=c('Ziziphi Spinosae Semen','Abri Herba'),type='latin')
}
