% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2t.R
\name{.m2t}
\alias{.m2t}
\title{how to make molecule to target}
\usage{
.m2t(x)
}
\arguments{
\item{x}{is the name of molecule}
}
\value{
target_name
}
\description{
how to make molecule to target
}
\examples{
.m2t(x='FER')
.m2t(x=c('FER','pyrene'))
}
